feat: "리뷰 화면 수정 및 무한스크롤 적용"

+++ molaeng_back/src/main/java/idle/molaeng_back/search/service/SearchServiceImpl.java
@@ -66,7 +66,6 @@ public class SearchServiceImpl implements SearchService {
     }
 
 
-
     @Override
     public SearchRecipeResDTO searchRecipeByCalory(Pageable pageable, long userId) {
         Slice<Recipe> recipeList = recipeRepository.findAll(pageable);
@@ -137,7 +136,7 @@ public class SearchServiceImpl implements SearchService {
             total += scoreCnt[i]*(i+1);
             size += scoreCnt[i];
         }
-        double avgScore = Math.round(((double)total*10/size)/10);
+        double avgScore = Math.round((total*10/size)/10.0);
         return avgScore;
     }
